Add files via upload
You are given a maze represented by a 2D array. The maze consists of walls represented by 'W', open
paths represented by 'O', the starting point represented by 'S', and the exit represented by 'E'. The
objective is to find a path from the starting point to the exit.
char maze[N][M] = {
{'S', 'O', 'O', 'W', 'W'},
{'O', 'W', 'O', 'O', 'W'},
{'O', 'O', 'O', 'W', 'O'},
{'W', 'W', 'O', 'W', 'O'},
{'W', 'W', 'O', 'E', 'W'}
};
Write a C program to traverse the maze of 5 x 5 and find a path from the starting point to the exit.
Constraints
- There will be at least one path from the starting point to the exit. In case of more than one path
then consider only one.
- You can visit only right and down direction.
- Output will be only indexes of correct path.